Team Web-site
This is the base of the website. Using the W3 to manipulate the layout with better control over creating css classes to be called. I used the same format changing the color scheme and some functions. On the (+) object that pulls-up a navigation side bar, I would like to add a like to different tables showing what each individual team members skills are.